Refactoring time blocks to remove redundancy
