+++ components/RemoverControl.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="sim-timeblock--remover">
+    <IconText icon="#icon--control--minus" icon-type="svg" />
+  </div>
+</template>
+
+<script>
+  import IconText from './IconText'
+
+  export default {
+    components: {
+      IconText,
+    },
+  }
+</script>
